refactor: rename types.py to schema.py to avoid stdlib conflict
