+++ whisperx/alignment.py
@@ -15,7 +15,7 @@ from transformers import Wav2Vec2ForCTC, Wav2Vec2Processor
 
 from whisperx.audio import SAMPLE_RATE, load_audio
 from whisperx.utils import interpolate_nans
-from whisperx.types import (
+from whisperx.schema import (
     AlignedTranscriptionResult,
     SingleSegment,
     SingleAlignedSegment,
