+++ whisperx/asr.py
@@ -12,7 +12,7 @@ from transformers import Pipeline
 from transformers.pipelines.pt_utils import PipelineIterator
 
 from whisperx.audio import N_SAMPLES, SAMPLE_RATE, load_audio, log_mel_spectrogram
-from whisperx.types import SingleSegment, TranscriptionResult
+from whisperx.schema import SingleSegment, TranscriptionResult
 from whisperx.vads import Vad, Silero, Pyannote
