+++ whisperx/schema.py
